fix(desktop): scope replay pagination to federation
