+++ apps/desktop/src/main/app-server/backend-registry.ts
@@ -40,6 +40,7 @@ import {
   formatManagedReviewOutput,
   parseManagedReviewOutput,
 } from "./managed-review";
+import { pageNormalizedReplay } from "./thread-replay-pagination";
 import {
   type AcpBackendId,
   buildAppendPinRank,
@@ -5851,69 +5852,6 @@ function mergeCodexDynamicToolSpecs(
   return [...functions, ...namespaces.values()];
 }
 
-function pageNormalizedReplay(
-  replay: AppServerThreadReplay,
-  options: {
-    before?: string;
-    includeTurns?: boolean;
-    limit?: number;
-  },
-): AppServerThreadReplay {
-  if (options.includeTurns === false) {
-    return {
-      ...replay,
-      entries: [],
-      messages: [],
-      lastUserMessage: undefined,
-      lastAssistantMessage: undefined,
-      pagination: {
-        supportsPagination: false,
-        hasPreviousPage: false,
-      },
-    };
-  }
-
-  if (options.limit === undefined && !options.before) {
-    return replay;
-  }
-
-  const endIndex = options.before
-    ? replay.entries.findIndex((entry) => entry.id === options.before)
-    : replay.entries.length;
-  const boundedEndIndex = endIndex >= 0 ? endIndex : replay.entries.length;
-  const limit =
-    options.limit === undefined ? undefined : Math.max(0, Math.floor(options.limit));
-  const startIndex = limit === undefined ? 0 : Math.max(0, boundedEndIndex - limit);
-  const entries = replay.entries.slice(startIndex, boundedEndIndex);
-  const messages = entries.flatMap((entry) =>
-    entry.type === "message"
-      ? [
-          {
-            id: entry.id,
-            role: entry.role,
-            text: entry.text,
-            ...(entry.parts ? { parts: entry.parts } : {}),
-            ...(entry.origin ? { origin: entry.origin } : {}),
-            ...(entry.createdAt ? { createdAt: entry.createdAt } : {}),
-          },
-        ]
-      : [],
-  );
-  const firstEntry = entries[0];
-  const hasPreviousPage = startIndex > 0;
-
-  return {
-    ...replay,
-    entries,
-    messages,
-    pagination: {
-      supportsPagination: true,
-      hasPreviousPage,
-      ...(hasPreviousPage && firstEntry ? { previousCursor: firstEntry.id } : {}),
-    },
-  };
-}
-
 function threadOrchestrationFailure(
   code: PwrAgentThreadOrchestrationErrorCode,
   message: string,
@@ -9454,7 +9392,7 @@ export class DesktopBackendRegistry {
       return await this.readAcpThread(request, backend);
     }
 
-    const backendReplay =
+    const replay =
       backend === "codex"
         ? await this.withCodexThreadClient(
             request.threadId,
@@ -9476,14 +9414,6 @@ export class DesktopBackendRegistry {
             before: request.before,
             limit: request.limit,
           });
-    // Some backends accept before/limit but return the complete transcript
-    // without pagination metadata. Bound that normalized replay locally so a
-    // long tool-heavy thread does not become one oversized renderer or
-    // federation payload.
-    const replay = backendReplay.pagination.supportsPagination
-      ? backendReplay
-      : pageNormalizedReplay(backendReplay, request);
-
     if (backend === "codex") {
       this.reconcileBackendCodexThreadStatus(request.threadId, replay.threadStatus);
     }
